app/requirements.txt
streamlit
requests
pulp
python-dotenv